fix: remove httpclient5 (#774)
All our services use the default version of HTTP client provided by spring-web. Here in study-server we introduced a new HTTP client provided by apache httpclient5. We have to double check if there was an important reason to do so. But to keep it harmonized between our services we prefer to remove it for now.

We witnessed unwanted side-effect of this client due to a pool of HTTP request that was very limited (only 5 per host, under load this would make all requests take minutes waiting for this pool!) and hard to configure (no spring auto-configuration)

Signed-off-by: Joris Mancini <joris.mancini_externe@rte-france.com>